requirements.txt
# HTTP & scraping
requests==2.32.3
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.48.0          # for JS-rendered pages (TGA search, etc.)

# Claude API
anthropic==0.40.0

# Scheduling
apscheduler==3.10.4

# Email
jinja2==3.1.4               # HTML email templates

# Storage
tinydb==4.8.0               # lightweight local JSON store (swap for Postgres later)

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2             # signal schema validation
tenacity==9.0.0             # retry logic for scrapers / API calls
